+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 - `cycle` now supports async iterables
+- `drop` now supports async iterables
 - `last` now supports async iterables
 - `size` now supports async iterables